(wip)refactor: put together the server actions of get data
