--- src/app/(routes)/users/[username]/(profile)/get-activities.ts
@@ -1,20 +1,0 @@
-import { annictApiClient } from '../../../../../lib/api/annict-rest/client'
-import { auth } from '../../../../../lib/auth'
-import { CACHE_TAGS } from '../../../../../lib/cache-tag'
-import type { User } from '../../../../../schemas/annict/users'
-
-export const getActivities = async (username: User['username']) => {
-  await auth()
-
-  const activitiesResult = await annictApiClient.getActivities(
-    { query: { filter_username: username, sort_id: 'desc' } },
-    { next: { tags: [CACHE_TAGS.USER(username), CACHE_TAGS.USER_ACTIVITY(username)] } },
-  )
-
-  if (activitiesResult.isErr()) {
-    console.error(`[/users/${username}] Failed to fetch activities:`, activitiesResult.error)
-    return null
-  }
-
-  return activitiesResult.value.activities
-}
