--- src/app/(routes)/works/[workId]/_layouts/get-work.ts
@@ -1,27 +1,0 @@
-import { notFound } from 'next/navigation'
-import { annictApiClient } from '../../../../../lib/api/annict-rest/client'
-import { auth } from '../../../../../lib/auth'
-import { CACHE_TAGS } from '../../../../../lib/cache-tag'
-import type { Work } from '../../../../../schemas/annict/works'
-
-export const getWork = async (workId: Work['id']) => {
-  await auth()
-
-  const worksResult = await annictApiClient.getWorks(
-    { query: { filter_ids: [workId], per_page: 1 } },
-    { next: { tags: [CACHE_TAGS.WORK(workId)] } },
-  )
-
-  if (worksResult.isErr()) {
-    console.error(`[/works/${workId}] Failed to fetch work:`, worksResult.error)
-    notFound()
-  }
-
-  const work = worksResult.value.works.at(0)
-
-  if (work === undefined || work === null) {
-    notFound()
-  }
-
-  return work
-}
