+++ src/app/actions/api/get/activities.ts
@@ -1,27 +1,44 @@
+'use server'
+
 import { annictApiClient } from '../../../../lib/api/annict-rest/client'
 import { auth } from '../../../../lib/auth'
 import { CACHE_TAGS } from '../../../../lib/cache-tag'
 import type { Activity } from '../../../../schemas/annict/activities'
 import type { User } from '../../../../schemas/annict/users'
 
-const groupByDate = (activities: Activity[]) => {
-  const currentDate = new Date()
-  const rangeDate = [...Array(30)].map((_, i) => {
-    const date = new Date(currentDate)
-    date.setUTCDate(date.getUTCDate() - i)
-    return date.toISOString().split('T')[0]
-  })
+export const getFollowingActivities = async (page = 1) => {
+  await auth()
 
-  return rangeDate.map((date) => {
-    const activitiesByDate = activities.filter((activity) =>
-      activity.created_at.startsWith(date),
-    ).length
+  const activitiesResult = await annictApiClient.getFollowingActivities(
+    { query: { per_page: 20, page, sort_id: 'desc' } },
+    { next: { tags: [CACHE_TAGS.ME, CACHE_TAGS.MY_ACTIVITY], revalidate: 60 } },
+  )
 
-    return { date, activities: activitiesByDate }
-  })
+  if (activitiesResult.isErr()) {
+    console.error('Failed to fetch following activities:', activitiesResult.error)
+    return null
+  }
+
+  return { data: activitiesResult.value.activities, next_page: activitiesResult.value.next_page }
 }
 
-export const getActivitiesPerDate = async (username: User['username']) => {
+export const getUserActivities = async (username: User['username'], page = 1) => {
+  await auth()
+
+  const activitiesResult = await annictApiClient.getActivities(
+    { query: { filter_username: username, page, per_page: 20, sort_id: 'desc' } },
+    { next: { tags: [CACHE_TAGS.USER(username), CACHE_TAGS.USER_ACTIVITY(username)] } },
+  )
+
+  if (activitiesResult.isErr()) {
+    console.error(`Failed to fetch user activities (${username}):`, activitiesResult.error)
+    return null
+  }
+
+  return { data: activitiesResult.value.activities, next_page: activitiesResult.value.next_page }
+}
+
+export const getUserActivityCountsPerDay = async (username: User['username']) => {
   await auth()
 
   const currentDate = new Date()
@@ -30,6 +47,22 @@ export const getActivitiesPerDate = async (username: User['username']) => {
     currentDate.getMonth() === 0 ? 11 : currentDate.getMonth() - 1,
     currentDate.getDate(),
   )
+
+  const groupByDay = (activities: Activity[]) => {
+    const rangeDate = [...Array(30)].map((_, i) => {
+      const date = new Date(currentDate)
+      date.setUTCDate(date.getUTCDate() - i)
+      return date.toISOString().split('T')[0]
+    })
+
+    return rangeDate.map((date) => {
+      const activitiesByDate = activities.filter((activity) =>
+        activity.created_at.startsWith(date),
+      ).length
+
+      return { date, activities: activitiesByDate }
+    })
+  }
 
   const getActivitiesImpl = async (page = 1): Promise<Activity[]> => {
     const PER_PAGE = 50
@@ -58,27 +91,13 @@ export const getActivitiesPerDate = async (username: User['username']) => {
   const filteredActivities = activities.filter(
     (activity) => new Date(activity.created_at) > lastMonthDate,
   )
-  const groupedActivities = groupByDate(filteredActivities)
+
+  const groupedActivities = groupByDay(filteredActivities)
 
   return groupedActivities
 }
 
-const groupByAction = (activities: Activity[]) => {
-  return activities.reduce(
-    (acc, activity) => {
-      acc[activity.action] += 1
-      return acc
-    },
-    {
-      create_record: 0,
-      create_multiple_records: 0,
-      create_review: 0,
-      create_status: 0,
-    },
-  )
-}
-
-export const getActivitiesPerAction = async (username: User['username']) => {
+export const getUserActivityCountsPerAction = async (username: User['username']) => {
   await auth()
 
   const currentDate = new Date()
@@ -87,6 +106,21 @@ export const getActivitiesPerAction = async (username: User['username']) => {
     currentDate.getMonth() === 0 ? 11 : currentDate.getMonth() - 1,
     currentDate.getDate(),
   )
+
+  const groupByAction = (activities: Activity[]) => {
+    return activities.reduce(
+      (acc, activity) => {
+        acc[activity.action] += 1
+        return acc
+      },
+      {
+        create_record: 0,
+        create_multiple_records: 0,
+        create_review: 0,
+        create_status: 0,
+      },
+    )
+  }
 
   const getActivitiesImpl = async (page = 1): Promise<Activity[]> => {
     const PER_PAGE = 50
