+++ src/app/actions/api/mutate/update-status.ts
@@ -1,10 +1,10 @@
 'use server'
 
 import { revalidateTag } from 'next/cache'
-import { annictApiClient } from '../../../lib/api/annict-rest/client'
-import { auth } from '../../../lib/auth'
-import { CACHE_TAGS } from '../../../lib/cache-tag'
-import type { Status } from '../../../schemas/annict/common'
+import { annictApiClient } from '../../../../lib/api/annict-rest/client'
+import { auth } from '../../../../lib/auth'
+import { CACHE_TAGS } from '../../../../lib/cache-tag'
+import type { Status } from '../../../../schemas/annict/common'
 
 export const updateStatus = async (workId: number, kind: Status) => {
   await auth()
